+++ packages/react-core/src/evm/providers/types.ts
@@ -41,8 +41,6 @@ export interface ThirdwebSDKProviderProps<TChains extends Chain[]>
     | (string & {})
     | Chain;
 
-  // api keys that can be passed
-  thirdwebApiKey?: string;
-  alchemyApiKey?: string;
-  infuraApiKey?: string;
+  // client Id for thirdweb services
+  clientId?: string;
 }