+++ packages/sdk/src/evm/constants/urls.ts
@@ -1,11 +1,10 @@
-import { DEFAULT_API_KEY } from "../../core/constants/urls";
 import type { ChainOrRpcUrl, NetworkInput } from "../core/types";
 import { isProvider, isSigner } from "../functions/getSignerAndProvider";
 import { StaticJsonRpcBatchProvider } from "../lib/static-batch-rpc";
 import type { SDKOptions, SDKOptionsOutput } from "../schema/sdk-options";
 import { SDKOptionsSchema } from "../schema/sdk-options";
 import type { ChainInfo } from "../schema/shared/ChainInfo";
-import { getChainRPC } from "@thirdweb-dev/chains";
+import { getValidChainRPCs } from "@thirdweb-dev/chains";
 import type { Chain } from "@thirdweb-dev/chains";
 import { providers } from "ethers";
 import type { Signer } from "ethers";
@@ -50,11 +49,11 @@ export function getChainProvider(
     // Resolve the chain id from the network, which could be a chain, chain name, or chain id
     chainId = getChainIdFromNetwork(network, options);
     // Attempt to get the RPC url from the map based on the chainId
-    rpcUrl = getChainRPC(rpcMap[chainId], {
-      thirdwebApiKey: options.thirdwebApiKey || DEFAULT_API_KEY,
-      infuraApiKey: options.infuraApiKey,
-      alchemyApiKey: options.alchemyApiKey,
-    });
+    rpcUrl = getValidChainRPCs(
+      rpcMap[chainId],
+      options.clientId,
+      options.secretKey,
+    )[0];
   } catch (e) {
     // no-op
   }
@@ -62,7 +61,7 @@ export function getChainProvider(
   // if we still don't have an url fall back to just using the chainId or slug in the rpc and try that
   if (!rpcUrl) {
     rpcUrl = `https://${chainId || network}.rpc.thirdweb.com/${
-      options.thirdwebApiKey || DEFAULT_API_KEY
+      options.clientId
     }`;
   }