+++ packages/sdk/src/evm/core/wallet/user-wallet.ts
@@ -8,10 +8,8 @@ import { AddressOrEns } from "../../schema/shared/AddressOrEnsSchema";
 import { Address } from "../../schema/shared/Address";
 import { SDKOptions } from "../../schema/sdk-options";
 import type { Amount, CurrencyValue } from "../../types/currency";
-import { ContractWrapper } from "../classes/contract-wrapper";
 import { RPCConnectionHandler } from "../classes/rpc-connection-handler";
 import { NetworkInput, TransactionResult } from "../types";
-import type { IERC20 } from "@thirdweb-dev/contracts-js";
 import ERC20Abi from "@thirdweb-dev/contracts-js/dist/abis/IERC20.json";
 import {
   type BigNumberish,
@@ -28,7 +26,9 @@ import type { BlockTag } from "@ethersproject/abstract-provider";
 import { fetchCurrencyValue } from "../../common/currency/fetchCurrencyValue";
 import { isNativeToken } from "../../common/currency/isNativeToken";
 import { normalizePriceValue } from "../../common/currency/normalizePriceValue";
-
+import type { IERC20 } from "@thirdweb-dev/contracts-js";
+import { ThirdwebStorage } from "@thirdweb-dev/storage";
+import { ContractWrapper } from "../classes/contract-wrapper";
 /**
  *
  * {@link UserWallet} events that you can subscribe to using `sdk.wallet.events`.
@@ -294,6 +294,7 @@ export class UserWallet {
       const localWallet = new UserWallet(
         new Wallet(LOCAL_NODE_PKEY, getChainProvider(chainId, this.options)),
         this.options,
+        // this.storage,
       );
       return localWallet.transfer(await this.getAddress(), amount);
     } else {
@@ -322,6 +323,11 @@ export class UserWallet {
       currencyAddress,
       ERC20Abi,
       this.options,
+      // TODO ideally we pass down storage from the constructor but causes problems in react-core
+      new ThirdwebStorage({
+        clientId: this.options?.clientId,
+        secretKey: this.options?.secretKey,
+      }),
     );
   }
 }